refactor: Rename webp.py and update import statements
